+++ commands/config_windows.go
@@ -11,6 +11,3 @@ func getDefaultConfigDirectory() string {
 		panic("Cannot get default config file location")
 	}
 }
-
-func userModeWarning(withRun bool) {
-}